# docs/api.yml
openapi: "3.1.0"
info:
  title: Grimoire
  version: "1.0.0"
  description: |
    Self-hosted TTRPG library manager.

    All endpoints except `/api/auth/status`, `/api/auth/setup`, and `/api/auth/login`
    require a valid JWT passed as `Authorization: Bearer <token>`, or as a `?token=`
    query parameter (required for browser-embedded images and file downloads).

    **Roles:**
    - `admin` — full access including user management and app settings
    - `gm` — read everything, edit metadata, rescan library, manage campaigns
    - `player` — read-only access
servers:
  - url: http://localhost:9481
    description: Local instance
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    tokenParam:
      type: apiKey
      in: query
      name: token
    apiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Token:
      type: object
      properties:
        id: { type: string, format: uuid }
        filename: { type: string }
        relative_path: { type: string }
        description: { type: string }
        tags: { type: array, items: { type: string } }
        is_explicit: { type: boolean }
        file_size: { type: integer }
        has_thumbnail: { type: boolean }
...
security:
  - bearerAuth: []
  - tokenParam: []
paths:
...
  /api/tokens:
    get:
      tags: [tokens]
      summary: List tokens
      parameters:
        - { in: query, name: limit, schema: { type: integer, default: 100 } }
        - { in: query, name: offset, schema: { type: integer, default: 0 } }
        - { in: query, name: tag, schema: { type: string } }
      responses:
        "200":
          description: Token list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Token" }
  /api/tokens/{id}:
    get:
      tags: [tokens]
      summary: Get a token
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Token detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Token" }
        "404": { description: Not found }
    patch:
      tags: [tokens]
      summary: Update token metadata
      description: GM or admin role required.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description: { type: string }
                tags: { type: array, items: { type: string } }
                is_explicit: { type: boolean }
      responses:
        "200": { description: Updated }
        "403": { description: GM or admin required }
        "404": { description: Not found }
...